+++ papers/catalog/hist_mag.py
@@ -26,7 +26,7 @@ from io import StringIO
 
 from sp_validation import run_joint_cat as sp_joint
 from sp_validation import util
-from sp_validation.basic import metacal
+from sp_validation.calibration import metacal
 from sp_validation import calibration
 import sp_validation.cat as cat
