+++ scripts/calibration/extract_info.py
@@ -37,7 +37,7 @@ from astropy.io import fits
 
 #from sp_validation.cat import *
 from sp_validation import cat as spv_cat
-from sp_validation.basic import *
+from sp_validation.calibration import metacal
 from sp_validation.calibration import *
 from sp_validation.galaxy import *
 from sp_validation.io import *
@@ -319,7 +319,7 @@ from cs_util.plots import plot_histograms
 from lenspack.geometry.projections.gnom import radec2xy
 from uncertainties import ufloat
 
-from sp_validation.basic import *
+from sp_validation.calibration import metacal
 from sp_validation.calibration import *
 from sp_validation.plot_style import *
 from sp_validation.plots import *
