+++ src/sp_validation/cat.py
@@ -26,7 +26,6 @@ from cs_util import cat
 
 from sp_validation import util
 from sp_validation import io
-from sp_validation import basic
 from sp_validation.survey import get_footprint
 from sp_validation import __version__, __name__
